Remove unnecessary '--rethrow-all' program option in tests.
In the circumstance where the default behavior was to (e.g.) skip the
event upon an exception throw, the '--rethrow-all' program option was
necessary to end the job.  However, the current default behavior is
that art will rethrow any exceptions the framework can specially
handle according to user configuration.

We should expand the exception-handling test suite to exercise
situations where users configure exception handling that differs from
the default behavior.